requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-mail==0.9.1
google-generativeai==0.7.2
python-dotenv==1.0.1
premailer==3.10.0
